minor: Move PreparedAccessPlan to same module as ParquetAccessPlan
